feature: add support conf

+++ src/Services/Projects/Projects.Application/ProjectMembers/DomainEventHandlers/ProjectMemberAddedDomainEventHandler.cs
@@ -16,8 +16,8 @@ public class ProjectMemberAddedDomainEventHandler : INotificationHandler<Project
     public Task Handle(ProjectMemberAddedDomainEvent notification, CancellationToken cancellationToken)
     {
         var @event = new ProjectMemberAddedEvent(notification.Member.Id, notification.Member.UserId,
-                                    notification.Member.UserEmail, notification.Member.PhotoUrl, (int)notification.Member.Type,
-                                     notification.Member.ProjectId, (int)notification.Member.InvitationStatus,
+                                    notification.Member.UserEmail, notification.Member.PhotoUrl, notification.Member.Type,
+                                     notification.Member.ProjectId, notification.Member.InvitationStatus,
                                      notification.Member.MotherProject.Name, notification.Member.MotherProject.IconUrl,
                                      notification.IsNewProjectCreator
                                      );
